+++ src/utils/checkAuth.ts
@@ -1,0 +1,5 @@
+export const checkAuth = async () => {
+  const token = localStorage.getItem("token");
+  if (token) return true;
+  return false;
+};
